Category was moved to the controller package.
